Normalize formatting and remove printf warnings
The source of this project is a total hack. Unified whitespacing is
a first step (no tabs, 4 spaces indent). Many of the xprintf:s had
problems with missing or extra arguments.

There was also warnings about %d for size_t. The 32/64 bit portable
solution is to use long format and cast to long, e.g "%lx" and
"(unsigned long)".

+++ src/yylex.c
@@ -83,7 +83,7 @@ char *placeIdent() {
 			assert(strlen(fn) < MAX_FILE_NAME_SIZE);
 			sprintf(mm, "%s:%d", simpleFileName(fn),cFile.lineNumber);
 			assert(strlen(mm) < MAX_HTML_REF_LEN);
-			sprintf(tt,"<A HREF=\"file://%s#%d\" %s=%d>%s</A>", fn, cFile.lineNumber, PPCA_LEN, strlen(mm), mm);
+			sprintf(tt, "<A HREF=\"file://%s#%d\" %s=%ld>%s</A>", fn, cFile.lineNumber, PPCA_LEN, (unsigned long)strlen(mm), mm);
 		} else {
 			sprintf(tt,"%s:%d ",simpleFileName(getRealFileNameStatic(cFile.fileName)),cFile.lineNumber); 
 		}
